Fix bip32 derivation (#94)
* Revert "BIP-32 Fix: Deactivate changes, to be reverted when finally merged"

This reverts commit c4d842ea

* BIP32 derivation merge

* Appkit develop branch

* Harden SigningSecrets.toBytes()

+++ common-jvm/src/main/java/org/ergoplatform/ApiServiceManager.kt
@@ -52,7 +52,7 @@ open class ApiServiceManager(
     override fun getUnconfirmedTransactions(limit: Int): Call<Transactions> =
         nodeTransactionsApi.getUnconfirmedTransactions(limit, 0)
 
-    override fun getExpectedWaitTime(fee: Long, txSize: Int): Call<Int> =
+    override fun getExpectedWaitTime(fee: Long, txSize: Int): Call<Long> =
         nodeTransactionsApi.getExpectedWaitTime(fee.toInt(), txSize)
 
     override fun getSuggestedFee(waitTime: Int, txSize: Int): Call<Int> =
